Use full Wikidata taxid list
The code is now using the full list of taxids connected to
Wikidata instead of querying them individually. This led to
a bug which makes the current Wikidata update fail to add a
taxon.

However, no errors are thrown when executing code.
Might be a problem in the way the Wikidata Integrator wrapper
functions call these taxids.