Add stub of tutorial to add pubmed id to Wikidata
As I will need to update Wikidata with the missing Cellosaurus references,
I will write the code for that.

Luckly, wikidata integrator has done already most of the work. To benefit
from it, I will first write a tutorial, and in the process, wrap the
functions for Cellosaurus. Maybe I will turn the functions into a nice
little module. Who knows?